css for nav and footer refactored into shared module
