--- src/Components/footer/Footer.module.css
@@ -1,26 +1,0 @@
-footer {
-  position: sticky;
-  bottom: 0;
-  background: linear-gradient(45deg, #007bff, #00d4ff);
-  color: white;
-  padding: 0.5rem 1rem;
-  text-align: center;
-  border-radius: 0 0 0.5rem 0.5rem;
-  box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
-  font-size: 0.9rem;
-}
-
-footer p {
-  margin: 0.5rem 0;
-  font-size: 0.9rem;
-}
-
-footer ul {
-  padding: 0;
-  margin: 0;
-}
-
-footer ul li {
-  display: inline;
-  margin: 0 0.5rem;
-}
