--- src/Components/navBar/NavBar.module.css
@@ -1,31 +1,0 @@
-/* Navbar */
-nav {
-  position: sticky;
-  top: 0;
-  background: linear-gradient(45deg, #007bff, #00d4ff);
-  color: white;
-  padding: 0.5rem;
-  text-align: center;
-  border-radius: 0.5rem 0.5rem 0 0;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-nav h1 {
-  font-size: 2.5rem;
-  font-weight: bold;
-}
-
-nav ul {
-  margin: 1rem 0 0.5rem 0;
-}
-
-nav ul li {
-  display: inline;
-  margin: 0 1rem;
-}
-
-nav ul li a {
-  padding: 0.5rem 1rem;
-  font-size: 1.1rem;
-  font-weight: bold;
-}
