+++ src/Components/sharedSyles/LayoutShared.module.css
@@ -1,0 +1,50 @@
+.shared_styles {
+  position: sticky;
+  background: linear-gradient(45deg, #007bff, #00d4ff);
+  color: white;
+  padding: 0.5rem;
+  text-align: center;
+}
+
+.shared_styles ul {
+  margin: 1rem 0 0.5rem 0;
+}
+
+.shared_styles ul li {
+  display: inline;
+  margin: 0 1rem;
+}
+
+.shared_styles ul li a {
+  padding: 0.5rem 1rem;
+  font-size: 1.1rem;
+  font-weight: bold;
+}
+/* Link buttons import own css ui custom link */
+/* nav */
+
+nav.nav {
+  top: 0;
+  border-radius: 1rem 1rem 0 0;
+}
+
+nav h1 {
+  font-size: 2.5rem;
+  font-weight: bold;
+}
+/* footer */
+footer.footer {
+  bottom: 0;
+  border-radius: 0 0 0.5rem 0.5rem;
+  font-size: 0.9rem;
+}
+
+footer.footer p {
+  margin: 0.5rem 0;
+  font-weight: bold;
+}
+
+footer.footer ul li a {
+  font-size: 0.625rem;
+  padding: 0.25rem 0.5rem;
+}
